+++ wallet/api/errors.go
@@ -151,18 +151,23 @@ var (
 	ErrSpecifyingNetworkAndLastBlockDataIsNotSupported    = errors.New("specifying a network and the last block data is not supported")
 	ErrSpecifyingNetworkAndNodeAddressIsNotSupported      = errors.New("specifying a network and a node address is not supported")
 	ErrSubmissionBlockHeightIsRequired                    = errors.New("the submission block height is required")
+	ErrTransactionCouldNotBeSentThroughSelectedNode       = errors.New("the transaction could not be sent through the selected node")
 	ErrTransactionFailed                                  = errors.New("the transaction failed")
 	ErrTransactionIsNotValidJSON                          = errors.New("the transaction is not valid JSON")
 	ErrTransactionIsRequired                              = errors.New("the transaction is required")
 	ErrUserCanceledTheRequest                             = errors.New("the user canceled the request")
 	ErrUserCloseTheConnection                             = errors.New("the user closed the connection")
-	ErrUserRejectedTheRequest                             = errors.New("the user rejected the request")
+	ErrUserRejectedAccessToKeys                           = errors.New("the user rejected the access to the keys")
+	ErrUserRejectedSendingOfTransaction                   = errors.New("the user rejected the sending of the transaction")
+	ErrUserRejectedSigningOfTransaction                   = errors.New("the user rejected the signing of the transaction")
+	ErrUserRejectedWalletConnection                       = errors.New("the user rejected the wallet connection")
 	ErrWalletAlreadyExists                                = errors.New("a wallet with the same name already exists")
 	ErrWalletDoesNotExist                                 = errors.New("the wallet does not exist")
 	ErrWalletIsLocked                                     = errors.New("the wallet is locked")
 	ErrWalletIsRequired                                   = errors.New("the wallet is required")
 	ErrWalletKeyDerivationVersionIsRequired               = errors.New("the wallet key derivation version is required")
 	ErrWrongPassphrase                                    = errors.New("wrong passphrase")
+	ErrTransactionBlockedBySpamRules                      = errors.New("the transaction will break the network's spam rules")
 )
 
 func applicationError(code jsonrpc.ErrorCode, err error) *jsonrpc.ErrorDetails {
@@ -191,7 +196,7 @@ func networkErrorFromTransactionError(err error) *jsonrpc.ErrorDetails {
 	txErr := types.TransactionError{}
 	isTxErr := errors.As(err, &txErr)
 	if !isTxErr {
-		return networkError(ErrorCodeNodeCommunicationFailed, ErrTransactionFailed)
+		return networkError(ErrorCodeNodeCommunicationFailed, ErrTransactionCouldNotBeSentThroughSelectedNode)
 	}
 
 	switch txErr.ABCICode {
@@ -234,8 +239,8 @@ func userCancellationError(err error) *jsonrpc.ErrorDetails {
 	return userError(ErrorCodeRequestHasBeenCanceledByUser, err)
 }
 
-func userRejectionError() *jsonrpc.ErrorDetails {
-	return userError(ErrorCodeRequestHasBeenRejected, ErrUserRejectedTheRequest)
+func userRejectionError(err error) *jsonrpc.ErrorDetails {
+	return userError(ErrorCodeRequestHasBeenRejected, err)
 }
 
 func applicationCancellationError(err error) *jsonrpc.ErrorDetails {