+++ wallet/api/node/types/types.go
@@ -23,6 +23,48 @@ type Statistics struct {
 	VegaTime    string
 }
 
+type SpamStatistics struct {
+	ChainID           string
+	EpochSeq          uint64
+	LastBlockHeight   uint64
+	Proposals         *SpamStatistic
+	Delegations       *SpamStatistic
+	Transfers         *SpamStatistic
+	NodeAnnouncements *SpamStatistic
+	IssuesSignatures  *SpamStatistic
+	Votes             *VoteSpamStatistics
+	PoW               *PoWStatistics
+}
+
+type SpamStatistic struct {
+	CountForEpoch uint64
+	MaxForEpoch   uint64
+	BannedUntil   *string
+}
+
+type VoteSpamStatistics struct {
+	Proposals   map[string]uint64
+	MaxForEpoch uint64
+	BannedUntil *string
+}
+
+type PoWStatistics struct {
+	PowBlockStates []PoWBlockState
+	PastBlocks     uint64
+	BannedUntil    *string
+}
+
+type PoWBlockState struct {
+	BlockHeight          uint64
+	BlockHash            string
+	TransactionsSeen     uint64
+	ExpectedDifficulty   *uint64
+	HashFunction         string
+	Difficulty           uint64
+	TxPerBlock           uint64
+	IncreasingDifficulty bool
+}
+
 type LastBlock struct {
 	ChainID                         string
 	BlockHeight                     uint64